+++ src/Products/PloneMeeting/tests/testWFAdaptations.py
@@ -5,6 +5,7 @@
 # GNU General Public License (GPL)
 #
 
+from AccessControl import Unauthorized
 from collective.behavior.internalnumber.browser.settings import get_settings
 from collective.behavior.internalnumber.browser.settings import set_settings
 from collective.contact.plonegroup.utils import get_all_suffixes
@@ -1550,6 +1551,7 @@ class testWFAdaptations(PloneMeetingTestCase):
     def _only_creator_may_delete_active(self):
         '''Tests while 'only_creator_may_delete' wfAdaptation is active.
            Only the 'MeetingMember' and the 'Manager' have the 'Delete objects' permission.'''
+        cfg = self.meetingConfig
         self.changeUser('pmCreator1')
         item = self.create('MeetingItem')
         self.assertEqual(item.query_state(), 'itemcreated')
@@ -1565,9 +1567,37 @@ class testWFAdaptations(PloneMeetingTestCase):
         self.changeUser('pmManager')
         # the MeetingManager can NOT delete
         self.failIf(self.hasPermission(DeleteObjects, item))
-        # God can delete too...
+        # Manager can delete
         self.changeUser('admin')
         self.failUnless(self.hasPermission(DeleteObjects, item))
+        # but annexes are still deletable by the users able to edit
+        # single delete
+        self.changeUser('pmManager')
+        annex = self.addAnnex(item)
+        decision_annex = self.addAnnex(item, relatedTo='item_decision')
+        self.assertEqual(get_annexes(item), [annex, decision_annex])
+        # not deletable by creator
+        self.changeUser('pmCreator1')
+        self.assertRaises(Unauthorized, item.restrictedTraverse('@@delete_givenuid'), annex.UID())
+        self.assertRaises(Unauthorized, item.restrictedTraverse('@@delete_givenuid'), decision_annex.UID())
+        # deletable by MeetingManager as item is "validated"
+        self.changeUser('pmManager')
+        self.portal.restrictedTraverse('@@delete_givenuid')(annex.UID())
+        self.portal.restrictedTraverse('@@delete_givenuid')(decision_annex.UID())
+        self.assertEqual(get_annexes(item), [])
+        # deletable using the delete batch action
+        annex = self.addAnnex(item)
+        decision_annex = self.addAnnex(item, relatedTo='item_decision')
+        self.assertEqual(get_annexes(item), [annex, decision_annex])
+        cfg.setEnabledAnnexesBatchActions(['delete'])
+        self.request['form.widgets.uids'] = u','.join([annex.UID(), decision_annex.UID()])
+        self.request.form['form.widgets.uids'] = self.request['form.widgets.uids']
+        self.request.form['ajax_load'] = 'dummy'
+        self.changeUser('pmManager')
+        form = item.restrictedTraverse('@@delete-batch-action')
+        form.update()
+        form.handleApply(form, None)
+        self.assertEqual(get_annexes(item), [])
 
     def test_pm_WFA_return_to_proposing_group(self):
         '''Test the workflowAdaptation 'return_to_proposing_group'.'''